Revise system headers / enforce c11 (#1695)
!!! Breaking Change !!!
----
All Sming based applications used to have a user_config.h file inside of the application `include/` directory. If your application user_config.h is the same as the one coming from the Sming then make sure to delete it .

If you don't want to delete it most probably the compilation will break with similar error message:
```
include/user_config.h:26:40: fatal error: espinc/c_types_compatible.h: No such file or directory
  #include <espinc/c_types_compatible.h>
```

Replace ` #include <espinc/c_types_compatible.h>` with `include <c_types.h>` to fix this issue.
-----

This PR includes the following changes:
* Standardise data types
* Use `uint32_t` in preference to `uint32`, etc.
* Arduino libraries use a mixture so still need to maintain definitions for those
* Build `C` files within modules to c11 standard (previously defaulted to `gnu90`)
* Move `espinc\c_types_compatible.h` -> `c_types.h`
* Simplify and tidy Esp8266 system headers
* Rename `espinc/c_types_compatible.h` as `c_types.h`, revise to use `stdint` and `stdbool` definitions
* Place location for `c_types.h` and `mem.h` ahead of SDK version in makefiles
* Move system definitions from `user_config.h` into `esp_systemapi.h` (exclude LWIP)
* Don't #include `user_config.h` from `esp_systemapi.h`
* Move missing memory API definitions out of `esp_systemapi.h` and into `mem.h`
* Simplify other headers
* Fix samples - use <esp_systemapi.h> to ensure correct c_types.h gets used
* Move attribute definitions and macros out of ctypes.h
* Add `WEAK_ATTR` macro to allow selective weak symbol support
* Not a C/C++ standard feature and not consistently applied on all platforms, so avoid where possible.

+++ Sming/Core/SPI.h
@@ -89,7 +89,7 @@ public:
 	 */
 	unsigned char transfer(unsigned char val) override
 	{
-		return transfer32((uint32)val, 8);
+		return transfer32(val, 8);
 	}
 
 	/** @brief read8() read a byte from SPI without setting up registers
@@ -105,7 +105,7 @@ public:
 	 * 	 USE WITH CARE !!
 	 *
 	 */
-	uint8 read8();
+	uint8_t read8();
 
 	/** @brief 	transfer16()
 	 * @param	short to send
@@ -122,10 +122,10 @@ public:
 	 */
 	unsigned short transfer16(unsigned short val) override
 	{
-		return transfer32((uint32)val, 16);
+		return transfer32(val, 16);
 	};
 
-	/** @brief 	transfer(uint8 *buffer, size_t numberBytes)
+	/** @brief 	transfer(uint8_t *buffer, size_t numberBytes)
 	 * @param	buffer in/out
 	 * @param	numberBytes lenght of buffer
 	 *
@@ -136,7 +136,7 @@ public:
 	 *
 	 * 		SPI.transfer(buffer, size)				: memory buffer of length size
 	 */
-	void transfer(uint8* buffer, size_t numberBytes) override;
+	void transfer(uint8_t* buffer, size_t numberBytes) override;
 
 private:
 	/** @brief transfer32()
@@ -150,7 +150,7 @@ private:
 	 * 		receivedVal = SPI.transfer(val)			: single byte
 	 * 		receivedVal16 = SPI.transfer16(val16)	: single short
 	 */
-	virtual uint32 transfer32(uint32 val, uint8 bits);
+	virtual uint32_t transfer32(uint32_t val, uint8_t bits);
 
 	// prepare/configure HSPI with settings
 	void prepare(SPISettings mySettings);
@@ -158,17 +158,17 @@ private:
 	/** @brief  spi_byte_order
 	 * private method used when applying SPISettings
 	 */
-	void spi_byte_order(uint8 byte_order);
+	void spi_byte_order(uint8_t byte_order);
 
 	/** @brief  spi_mode
 	 * private method used when applying SPISettings
 	 */
-	void spi_mode(uint8 mode);
+	void spi_mode(uint8_t mode);
 
 	/** @brief  setClock
 	 * private method used when applying SPISettings
 	 */
-	void setClock(uint8 prediv, uint8 cntdiv);
+	void setClock(uint8_t prediv, uint8_t cntdiv);
 
 	/** @brief  setClock
 	 * private method used when applying SPISettings
